Task 3 start
It runs, but I can't get it to run as I want it. This is a daemon, and
ideally it runs apart from bash, and you can run whatever you want on
bash and it'll do it's thing. After I get it worked out, I might write a
program to encapsulate the daemon, its tests, and killing the daemon.